more multi os support

+++ master.go
@@ -1,10 +1,10 @@
 package main
 
 import (
-	"fmt"
+	"bytes"
 	"io/ioutil"
 	"os"
-	"runtime/debug"
+	"os/exec"
 
 	"golang.org/x/sys/unix"
 )
@@ -12,12 +12,9 @@ import (
 func master() error {
 	var err error
 
-	if err = ioutil.WriteFile("/proc/self/oom_score_adj", []byte("-1000"), 0644); err != nil {
-		l.Debug(err.Error())
-	}
-
-	if err = ioutil.WriteFile("/proc/self/oom_adj", []byte("-17"), 0644); err != nil {
-		l.Debug(err.Error())
+	master := true
+	if os.Getenv("master") == "false" {
+		master = false
 	}
 
 	if err = os.Chdir("/"); err != nil {
@@ -26,33 +23,55 @@ func master() error {
 
 	unix.Umask(0)
 
-	if err = unix.Setpgid(0, 0); err != nil {
-		l.Debug(err.Error())
+	if master {
+		if err = ioutil.WriteFile("/proc/self/oom_score_adj", []byte("-1000"), 0644); err != nil {
+			l.Debug(err.Error())
+		}
+
+		if err = ioutil.WriteFile("/proc/self/oom_adj", []byte("-17"), 0644); err != nil {
+			l.Debug(err.Error())
+		}
+
+		if err = unix.Setpgid(0, 0); err != nil {
+			l.Debug(err.Error())
+		}
+
+		if _, err = unix.Setsid(); err != nil {
+			l.Debug(err.Error())
+		}
+
+		for _, pid := range getPids("qemu-ga", true) {
+			unix.Kill(pid, unix.SIGTERM)
+		}
 	}
 
-	if _, err = unix.Setsid(); err != nil {
-		l.Debug(err.Error())
-	}
-
-	for _, pid := range getPids("qemu-ga", true) {
-		unix.Kill(pid, unix.SIGTERM)
-	}
 	/*
 		syscall.Close(0)
 		syscall.Close(1)
 		syscall.Close(2)
 	*/
 
-	defer func() {
-		if err := recover(); err != nil {
-			l.Error(fmt.Sprintf("%v %s", err, debug.Stack()))
+	if master {
+		stdOut := bytes.NewBuffer(nil)
+		stdErr := bytes.NewBuffer(nil)
+		for {
+			cmd := exec.Command("qemu-ga")
+			cmd.Dir = "/"
+			cmd.Env = append(cmd.Env, "master", "false")
+			cmd.Stdin = nil
+			cmd.Stdout = stdOut
+			cmd.Stderr = stdErr
+			cmd.ExtraFiles = nil
+			cmd.Run()
+			stdErr.Reset()
+			stdOut.Reset()
 		}
-	}()
-
-	for {
-		if err = slave(); err != nil {
-			l.Error(err.Error())
+	} else {
+		for {
+			if err = slave(); err != nil {
+				l.Error(err.Error())
+			}
 		}
 	}
-	//	return nil
+
 }
